Ongoing with fast mixed solver
The best way to compute volumetric integrals is to transfer all
geometric and space functions on integration points.

+++ Projects/iRMS/TRMSpaceOdissey.h
@@ -116,7 +116,7 @@ public:
     void PrintGeometry();
     
     /** @brief Create a reservoir-box geometry */
-    void CreateGeometricBoxMesh(TPZManVector<int,2> dx, TPZManVector<int,2> dy, TPZManVector<int,2> dz);
+    void CreateGeometricBoxMesh(TPZManVector<REAL,2> dx, TPZManVector<REAL,2> dy, TPZManVector<REAL,2> dz);
     
     /** @brief Parametric function that computes elements in the x direction */
     static  void ParametricfunctionX(const TPZVec<STATE> &par, TPZVec<STATE> &X);
